Add authentication handler
This patch adds an HTTP handler that authenticates users based on the
JWT token included in the `Authorization` HTTP header.

By default the handler will accept all requests and map them to a
`guest` user.

In order to actually perform authentication the servers will need to be
executed with the `--auth-jwks-*` options. For example, to run the
deployment manager server authenticating with the Red Hat external SSO:

```
$ ./oran-o2ims start deployment-manager-server \
--authn-jwks-url=https://sso.redhat.com/auth/realms/redhat-external/protocol/openid-connect/certs
```

Another example, this one to run the server as a pod inside a Kubernetes
cluster and authenticate with the cluster service accounts:

```
$ ./oran-o2ims start deployment-manager-server \
--authn-jwks-url=https://kubernetes/openid/v1/jwks \
--authn-jwks-token-file=/run/secrets/kubernetes.io/serviceaccount/token \
--authn-jwks-ca-file=/run/secrets/kubernetes.io/serviceaccount/ca.crt
```

Related: https://issues.redhat.com/browse/MGMT-16123
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>