Add development reverse proxy
This patch adds a `proxy.sh` script that starts the servers and the
reverse proxy in the local machine. The reverse proxy is Envoy, and the
configuration is in the `proxy.yaml` file.

To use it run the `run.sh` script. It will start the metadata server
listening in port 8000, the deployment manager server listening in port
80001 and the reverse proxy listening in port 9000. Then, for example,
go to port 9000 to get the list API versions served by the metadata
server:

  http://localhost:9000/O2ims_infrastructureInventory/api_versions

Or to get the the list of deployment managers server by the deployment
manager server:

  http://localhost:9000/O2ims_infrastructureInventory/v1/deploymentManagers

To finish, type Ctrl+C and the proxy and the servers will be killed.

The administration port of the proxy is 9001. There you can see Envoy
cofguration details and metrics.

This reverse proxy configuration is intended for development
environments only, but will be the seed for the production
configuration.

Related: https://issues.redhat.com/browse/MGMT-16113
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>