MGMT-16108: introduce the Resource server
Added an initial implementation for the resource server.
The server is serving the 'resourcePools' endpoint which
fetches data from the search query API and converts
the results into O2 objects.

In a nutshell, the mapping is done as follows:
* Node object from search API --> O2 Resource object
* Cluster object from search API --> O2 ResourcePool object

As some objects/properties don't have direct mapping,
for now this PR does *not* include the following:
* Handling for ResourceType objects
* Mapping for these properties:
  * ResourcePool.location
  * ResourcePool.globalLocationID
  * ResourcePool.description
  * Resource.resourceTypeID