Simplify deployment manager handler (#27)
This patch changes the implementation of the deployment manager service
so that it has one single handler for the collection and for individual
items. That is convenient because the code for both things is very
similar, in particular the mapping code is identical. In order to do
that simplification it is necessary to ensure that the methods of the
`CollectionHandler` and `ObjectHandler` have different names, as
otherwise it isn't possible to implement both in the same object.
Currently both are named `Get`. This patch renames that method to
`List`, merges the deployment manager code into one single handler, and
adds tests to verify that it works correctly.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>