MGMT-16351: Add support for configured extensions
Currently the population of the _extensions_ field is completely
hardcoded. This patch adds support for populating it in a configurable
way. A new `--extensions` command line option is added to the deployment
manager server. The value will be a jq script that calculates extension
fields. For example, if the managed cluster has a `country` label the
following command line option will configure the server to add a
corresponding `country` extension:

```
--extensions={
  "country": .metadata.labels["country"]
}
```

For a cluster where the value of the `contry` label is `ES` that will
result in the following extensions:

```json
{
  "country": "ES"
}
```

Related: https://issues.redhat.com/browse/MGMT-16351
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>