Fix projection with multiple fields (#38)
Currently when a projector contains multiple nested fields an extra
field will be added with the name of the container. For example, if
the projector contains `a/b,a/c` the result will be:

```
{
  "a" : {
    "b": "B"
    "c": "C",
    "a": "C"
  }
}
```

Note the extra `a.a` field, with value `C`. That shouldn't be there, it
is an artifact of the projector evaluator. This patch fixes that issue.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>