MGMT-16307: Optimize GraphQL search using filters (#41)
Optimized the filtering of search results in the resource server.
I.e. instead of filtering the items after fetching a response,
used the filter functionality of GraphQL search query API.
This change includes the required mapping from 'filter' query param
(as defined in the NFV spec) into SearchFilter as used in GraphQL.
For any unsupported search operator by GraphQL, we fallback to
the regular selector filtering.