MGMT-16359: support configured extensions in resource server
Similarly to the support add in deployment manager server,
this PR is adding a similar functionality to the resource server.
The main difference between the implementations is the parsing
logic of the objects retrieved from search API.

I.e. since each 'label' property is a single string consists
of key=value pairs, the lables should be parsed before
evaluating the expression with jq.
Hence, added a custom jq function for parsing: `parse_labels`.

For example:
```
--extensions={
  "openshift_version": .label|parse_labels|.openshiftVersion,
  "cluster_set": .label|parse_labels|.\"cluster.open-cluster-management.io/clusterset\",
}
```