Add support for regular hub (#49)
Currently the deployment manager server expects to have an ACM global
hub as the backend. This patch adds support for using a regular ACM hub.
The new `--backend-type=global-hub|regular-hub` command line flag will
be used to select the type of hub. When using a global hub the
deployment manager server will use the non-kubernetes API of the hub.
When using a regular hub it will use the `ManagedCluster` Kubernetes
API.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>