Add POST support (#66)
This patch adds support for the `POST` HTTP method. The adapter will now
accept the `POST` method and call the `Add` method defined in the
`AddHandler` interface. An implementation of this interface will receive
the request body already parsed, and should add the object to the
collection and return it. For example, a simple implementation that
stores the object in a map in memory could look like this:

```go
func (h *MyHandler) Add(ctx context.Context, request *AddRequest) (respone *AddResponse, err error) {
	h.storeLock.Lock()
	defer h.storeLock.Unlock()
	h.storeMap[request.Variables[0]] = request.Object
	response := &AddResponse{
		Object: request.Object,
	}
	return
}
```

In order to make naming of the interfaces more consistent this patch
also renames `ObjectHandler` to `GetHandler` and `CollectionHandler` to
`ListHandler`. That way all interfaces are named after the operation
they represent.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>