Add defaults for deployment manager backend URL and token
Currently the backend URL and token don't have defaults. That means that
the user has to explicitly set them in the custom resource that
represents the O2 IMS deployment. That is reasonable for the URL, but
not for the token, because it is the token of a service account that is
created by the operator. That means that the user would have to create
the service account in advance, wait for the token to be created and
then read it and update the custom resource. That isn't practical. To
avoid that this patch changes the operator so that the token is the
default URL of the Kubernetes API server and the token is taken from the
`/run/secrets/kubernetes.io/serviceaccount/token` file.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
Co-authored-by: Irina Mihai <imihai@redhat.com>