Use assisted installer admin kubeconfig if exists
Currently the deployment manager service uses the admin kubeconfig that
was used to register the cluster with ACM. But this will not exist if
the cluster was created with assisted installer. To avoid that problem
this patch changes the service so that it first check if the secret used
by the assisted installer exists. If it does exist then it will use it.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>