Add reconciler task
This patch adds a new `reconcilerTask` that contains the logic and data
needed to perform the reconciliation process. It contains, for example,
the state of the object being reconciled. This reduces the number of
parameters that need to be passed to methods, as some of them are
available as fields of the receiver.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>