Move controller manager command to internal
Currently the code of the command that starts the operator is in a
separate `main.go` file inside the `cmd` directory. This is different to
all the other commands that start things, as they are sub-commands of
the `oran-oims start` command and their code lives in the `internal/cmd`
package. In addition that operator command also uses a different logging
framework than the rest of the project. In order to improve consistency
this patch changes that command to use the same infrastructure than the
rest of the project, including logging. The command will now be
`oran-o2ims start controller-manager`.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>