Reduce numer of parameters of `CreateK8sCR` (#85)
Currently this function has nine parameters. Two of them are the name
and namespace of the object, but the object itself is also passed.
Assuming that the `Name` and `Namespace` fields of the object will be
populated it is not necessary to have additional parameters for that.
This patch removes those parameters.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>