Remove `oldObject` parameter from `CreateK8sCR` (#87)
Currently the `CreateK8sCR` function receives an `oldObject` parameter
that is needed because the function doesn't know how to create an empty
object of the same type of the object that is going to be created. But
that can be solved using reflection, something like this:

```go
// Create an empty object of the same type of the new object. We will use it to store the
// fetch the current state.
objectType := reflect.TypeOf(newObject).Elem()
oldObject := reflect.New(objectType).Interface().(client.Object)
```

This patch does that and removes the `oldObject` parameter, which
simplifies the code that calls the `CreateK8sCR` function.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>