Support creation of non-namespaced objects in `CreateK8sCR` (#86)
Kubernetes forbids the creation of object in one namespace with owners
in another namespace, or the creation of non-namespaced objects with
namespaced owners. But the `CreateK8sCR` function doesn't support. We
will need that in order to create cluster roles and cluster role
bindings. This patch changes the function so that it will not try to set
an owner if the namespace of the object and the namespace of the owner
are different.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>