Grant deployment manager server account permissions to read managed clusters and secrets (#82)
* Update generated files

This patch regenerates manifest files generated by `make generate` and
`make bundle`.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>

* Grant deployment manager server required permissions

The service account that is used to run the deployment manager server
needs permissions to read `ManagedCluster` objects and assisted
installer admin kubeconfig secrets. This patch changes the operator so
that it creates the corresponding cluster roles and cluster role
bindings. In order to do so the operator needs to also have those
permissions, as Kubernetes forbids granting more permissions than the
service account granting them already has. So this patch also adds those
permissions to the service account that is used to run the operator.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>

---------

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>