Make image configurable
Currently the container image used by the servers is hardcoded in the Go
source code. This complicates development because when the operator is
deployed it isn't possible to use a custom image. To simplify that this
patch adds a new `--image` command line flag to the `start
controller-manager` command. The default will be
`quay.io/openshift-kni/oran-o2ims:latest`, but it will be possible to
change it in the the development environment, in two different ways:

1. Changing the `IMG` make variable, for example:

  ```
  $ make IMG=quay.io/myuser/myimage:123 build deploy
  ```

2. Using the `spec.image` field of the custom resource.

The patch also changes The kustomization files used to generate tha
operator manifests so that the selected image is automatically passed to
the `--image` command line flag.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>