Add support for paging
This patch adds support for the `nextpage_opaque_marker` parameter as
defined in section 5.4 of ETSI GS NFV-SOL 013. A new `NextPageMarker`
field is added to the `ListRequest` and `ListResponse`. The adapter
takes care of encoding and encrypting it, and the handler only needs
to read it from the request or write it to the response. The type of
the field is `[]byte` and the handler that uses it is responsible for
encoding and decoding it if needed.

Related: https://www.etsi.org/deliver/etsi_gs/NFV-SOL/001_099/013/03.05.01_60/gs_NFV-SOL013v030501p.pdf
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>