Use logging context where possible (#95)
Currently whe use the `Info`, `Error` and `Debug` methods of the
`slog.Logger` package, which don't receive a context. There are
`InfoContext`, `ErrorContext` and `DebugContext` that receive a context
as a parameter and make it possible to later include context information
in the generated log messages. This patch changes our code to use those
variants, so that we will eventually be able to use that context
information.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>