Keep status conditions in order of their original popluation logic (#196)
Currently, if the condition status is changed, it always be added in
the last condition in the list, which causes the confusion. This update
will keep the original order of the conditions as they were initially
popluated according to a specific logic.

Start handling cluster provisioning only when the hwProvisioned condition
is available (also consider provisioning completed when the hardware plugin is ready),
and start handling policy configuration only when the clusterProvisioned
condition is available.