Change the rendered ClusterInstance's type to siteconfig.ClusterInstance (#195)
The ClusterInstance template is explicitly defined and delivered within the IMS operator.
When new fields are available in the Siteconfig operator, we will need to update the
ClusterInstance template and upversion the IMS operator to support the new fields.

Given this, we could switch to rendering the ClusterInstance into the ClusterInstance
struct type imported from the siteconfig package instead of using the unstructured type,
which will simplify data manipulation. It's safe to directly manipulate with the
ClusterInstance data, as the non-empty parameters have been verified in the ClusterInstance
template and validated with dry-run previously.

When we need to update the ClusterInstance template to support new fields, the imported
siteconfig package should be upversioned as well.