MGMT-18976: Use kube-rbac-proxy for authentication and authorization (#211)
* MGMT-18976: Use kube-rbac-proxy for authentication and authorization

This replaces the current authentication and authorization code by
leveraging the kube-rbac-proxy.  This reduces the amount of code we
must support and allows us to support the standard Kubernetes RBAC
mechanism for authorizing users to use the O2IMS API endpoints.

Signed-off-by: Allain Legacy <alegacy@redhat.com>

* MGMT-18976: Update how rbac proxy image is injected

This creates a new kustomization replacement to copy the rbac proxy
image into the manager's environment variables so that it can ultimately
be referenced whenever starting a new server.

Signed-off-by: Allain Legacy <alegacy@redhat.com>

---------

Signed-off-by: Allain Legacy <alegacy@redhat.com>