Change operation type to update for services (#224)
The fact that we were using a "patch" operation on services is leading
to some errors in the reconcile loop.  The "patch" should be reserved
to use cases where we read an object from the system, make some changes
to our local copy, and then want to "patch" it using the API to only
send those fields that have changed.  What we were doing was creating
a brand new object and attempting to "patch" that against what the
running system already had which lead to trying to delete several
runtime only attributes which I believe is what was leading to the
errors.

Signed-off-by: Allain Legacy <alegacy@redhat.com>