Add Interface Label to clusterinstance-defaults
This update adds a label for each interface specified in the
default cluster instance ConfigMap. The label is used to match the
corresponding interface label returned by the hardware manager. Upon
finding a match, the returned MAC address is assigned to the respective
interface in the rendered cluster instance.

Additionally, this update moves hardware provisioning-specific utility
functions to a separate file for improved organization and clarity.

Signed-off-by: Tao Liu <tali@redhat.com>