Add client certificate support for mTLS (#283)
This adds the ability to add a client certificate to the SMO and OAuth
configuration attributes so that mTLS can be enabled towards both of
those servers.  How that certificate is created and managed is a
separate concern and for now is left as a manual exercise for the
end user.

Signed-off-by: Allain Legacy <alegacy@redhat.com>