MGMT-18997: Load default Inventory CR on startup (#292)
* MGMT-18997: Load default Inventory CR on startup

This changes the operator startup sequence so that a default Inventory
CR is now created when the operator first loads.  This eliminates the
need for an end user to create an Inventory CR just for the purpose
of starting the servers and making the API endpoints accessible.  The
end user still must edit the default Inventory CR to provision the
Global Cloud ID and the SMO configuration attributes.  The end user
may also edit the Inventory CR to override attributes such as the
Ingress URL and the backend URLs if the operator is running in a non-
standard environment.

Signed-off-by: Allain Legacy <alegacy@redhat.com>

* MGMT-18997: Address PR comments

---------

Signed-off-by: Allain Legacy <alegacy@redhat.com>