CNF-15167: Fix RBAC role for searching managed data
The previous PR for this Jira addressed the managed data use case, but
inadvertently disabled the local cluster use case.  This change
reverts the first since the allManagedData RBAC rule seems to make it
such that the service account can only see managed data and not any
local data.  To enable querying the managed data we require the ability
to create managedclusterviews.

Signed-off-by: Allain Legacy <alegacy@redhat.com>