MGMT-18997: Get default domain from status rather than spec (#308)
This changes how we acquire the default domain so that we look at the
ingress controller's Status field rather than its Spec field since in
some cases it may not be explicitly set in the Spec which causes it to
go get it from another source.  Ultimately, we do not care where it gets
it from so we look at its final decision which is stored in the Status
field.

Signed-off-by: Allain Legacy <alegacy@redhat.com>