CNF-15750: Cleanup inventory API and code generate server (#332)
* CNF-15750: Cleanup inventory API and code generate server

This commit is the beginning of aligning the inventory (resource) API
endpoints with the new implementation of the alarm server.  In order
to reduce boilerplate code the inventory (resource) server structure
will now be code generated from the openapi yaml.  Since certain API
components (e.g.,  parameter definitions) are common to both the alarm
and inventory components they are split out into a "common" package.

Signed-off-by: Allain Legacy <alegacy@redhat.com>

* CNF-15750: Address PR comments

Signed-off-by: Allain Legacy <alegacy@redhat.com>

---------

Signed-off-by: Allain Legacy <alegacy@redhat.com>