CNF-14144: Add migration init container for resource server
This commit adds the init container for the resource server.  The
container runs prior to starting the resource server container so that
its database schema is created and migrated up to the latest schema
version.  The init container is run any time the resource server pod
is recreated, and if there is a transient failure in the init due to
a race condition in bringing up the database server it is automatically
retried until it succeeds.

Signed-off-by: Allain Legacy <alegacy@redhat.com>