CNF-14955: Adding postgres resources to controller (#336)
* CNF-14955: Adding postgres resources to controller

This adds the ability to manage the Kubernetes resources related to the
Postgres database.  This is a first iteration which does not include
the management of the PV/PVC for the persistent data.  It handles only
the establishment of the deployment and the user credentials.

Signed-off-by: Allain Legacy <alegacy@redhat.com>

* CNF-14955: Remove postgres k8s resources

Removes the manual testing resources used to deploy postgres manually
prior to the operator being able to do it autonomously.

Signed-off-by: Allain Legacy <alegacy@redhat.com>

* CNF-14955: Adding postgres resources to controller

Addressing PR comments

Signed-off-by: Allain Legacy <alegacy@redhat.com>

* CNF-14955: Adding postgres error conditions

Signed-off-by: Allain Legacy <alegacy@redhat.com>

* CNF-14955: Fix status condition update

Signed-off-by: Allain Legacy <alegacy@redhat.com>

---------

Signed-off-by: Allain Legacy <alegacy@redhat.com>