Lookup search api URL dynamically (#347)
* Lookup search api URL dynamically

Rather than hardcode the name and namespace of the search-api service
this commit searches for it using its well-known search labels.  As is
currently the case, if a user provides an override on the Inventory CR
then we use that rather than searching for it.

Signed-off-by: Allain Legacy <alegacy@redhat.com>

* Allow override dynamic fields

This commit changes the inventory reconciler so that an end user can
override any of the dynamic fields (i.e., ingress domain, search-api)
after the system has already been initialized.

Signed-off-by: Allain Legacy <alegacy@redhat.com>

---------

Signed-off-by: Allain Legacy <alegacy@redhat.com>