Align API path endpoints with specification (#353)
This aligns the newly defined openapi specification with the naming
used in the specification document.  The specification document
inconsistently uses capitalization and underscores in the naming of the
API endpoint paths, but based on the conformance tests and the
specific sections of the documents that deal with each of the
individual endpoints it seems clear that the intent is for them to
start with "o2ims-" rather than "O2ims_".

Signed-off-by: Allain Legacy <alegacy@redhat.com>