CNF-15750: Add field and record filtering to new API servers
This ports the existing include/exclude field filter as well as the
record filter from the existing API implementation to the new API
implementation for both the alarm and resource servers.  See the common
openapi.yaml file for more information on the functioning of the query
parameters for 'fields', 'exclude_fields', and 'filter' parameters.

The 'filter' functionality should eventually be integrated directly into
the database queries to limit the number of results directly at the
source rather than post-processing the result which is far more
expensive. The current approach is a temporary solution to achieve
parity with the existing API implementation with the least amount of
cost.

Signed-off-by: Allain Legacy <alegacy@redhat.com>