CNF-15570: Add inventory data collector
This adds the first pass at implementing the data collector for the
inventory data.  This initial implementation synchronizes the
resource pool, resource type, and resource objects from the ACM and
creates corresponding data change event records to track when any of
these objects are created or updated.  A future commit will address the
delete use case as well as the other objects (e.g., deployment manager).

Signed-off-by: Allain Legacy <alegacy@redhat.com>