CNF-15570: Add notification support (#407)
* CNF-15570: Add notification support

This adds support for sending notifications to a subscriber whenever
a data change event occurs.  This applies to all the supported
inventory object (i.e., resource types, resource pools, resources,
and deployment managers).

If the subscriber is not reachable then a number of retries are
attempted.  If at the end of the retries the subscriber is still not
reachable then the event notification is retired and the notifier moves
on to the next event in the list.  This behaviour should likely be
changed so that if the subscriber is unreachable for a set period of
time then the subscription should be put in an 'offline' state and
should ignore new events for some period of time to avoid excess error
logs.   This is left for a future commit.

Signed-off-by: Allain Legacy <alegacy@redhat.com>

* CNF-15570: Address PR comments

Signed-off-by: Allain Legacy <alegacy@redhat.com>

---------

Signed-off-by: Allain Legacy <alegacy@redhat.com>