CNF-16217: Add support for O2IMS Cluster namespace
This introduces the new "cluster" namespace of the O2IMS API which is
meant to represent CaaS resources in contrast with the existing
"inventory" namespace which is meant to represent h/w type resources.
Our current implementation doesn't 100% conform to the newly expressed
content in the upcoming document version.  This commit adds the new
object models to align with the document.

This is a first step towards having the ability to report alarms against
a cluster.  The next step will be to track alarm dictionaries against
these objects and to exchange information with the alarm server.

Signed-off-by: Allain Legacy <alegacy@redhat.com>