Implement Alarm Service Configuration endpoints
This implements the alarm service configuration endpoints.
The specification defines GET, PUT (only update, not create)
and PATCH operations. Therefore, we create a default
service configuration at startup. The spec also mentions that
the retention period has a minimum value which is not explicitly
defined. We assume 10 and 1 as default and minimum retention periods
respectively. The values represent days.

Signed-off-by: Marcelo Guerrero <marguerr@redhat.com>