Fix notifier deadlock on new events
The design of the notifier and notifier worker interactions was flawed
in that it allowed the possibility of a deadlock whenever both the
worker's work channel and notifier's completion channel buffers were
full.  The notifier cannot send new work to the worker, and the
worker gets stuck waiting for the notifier to acknowledge receipt of the
completion event.

This is being mitigated by having the notifier enqueue work onto the
worker's work queue directly, and only signaling the work channel upon
adding the first event onto the queue.  The worker will independently
process the queue and report completion down to the notifier.  This
should eliminate the possibility that the completion channel fills up
and the worker blocking indefinitely.

Signed-off-by: Allain Legacy <alegacy@redhat.com>