CNF-16243: Transition to new resource server (#448)
* CNF-16243: Transition to new resource server

This commit transitions to using the new implementation of the resource
server.  This enables the use of notifications to subscribers when
inventory data changes.  For now this is limited to polling of inventory
at 10 minute intervals.  In the future this will evolve to an
edge-driven approach depending on whether the hardware manager plugin
supports this type of update mechanism.

Signed-off-by: Allain Legacy <alegacy@redhat.com>

* CNF-16243: Obsolete servers after transition to new resource server

This removes the metadata and deployment manager servers since they are
now obsolete.  Their functionality has been moved to the new resource
server implementation.

---------

Signed-off-by: Allain Legacy <alegacy@redhat.com>