Fix notifier related hangs on shutdown (#455)
The input channels to the notifier for subscription or notification
events assume that the write to the channel will always succeed.  Since
we used buffered channels, and we know that there are use cases where
many events may be in-flight simultaneously, it is conceivable that the
notifier back-pressures back to the source of the events.  If this
happens, then the event source blocks until the channel becomes
available to receive a new message.

If this condition is present while the process is shut down, then the
source go routine will hang waiting to send to the channel(s).  By
incorporating the channel write into a select that also considers the
state of the context; then we are able to abort a blocked send if the
source go routine is canceled.

Signed-off-by: Allain Legacy <alegacy@redhat.com>