CNF-16247: Add support for in-cluster subscription endpoints
This adds the ability to re-use our subscription mechanism between
services rather than solely supporting OAuth public endpoints (e.g.,
SMO).  The type of client created is based on the callback URL provided
in the subscription.  If it contains "svc.cluster.local" then it is
assumed to be in-cluster and uses the service account token rather than
an OAuth token.

Signed-off-by: Allain Legacy <alegacy@redhat.com>