CNF-16245: Add edge driven cluster updates with Reflector (#460)
* CNF-16245: Add edge driven cluster updates with Reflector

This adds the ability to watch for changes to ManagedCluster and Agent
objects so that we can dynamically update our persistent storage rather
than have to poll at fixed intervals.  The watch is implemented using a
Reflector and supplying an adapter layer that receives cache update
events and feeds them to the Collector which then updates our storage
and emits async notification events to subscribers as needed.

Signed-off-by: Allain Legacy <alegacy@redhat.com>

* CNF-16245: Address PR comments

Signed-off-by: Allain Legacy <alegacy@redhat.com>

---------

Signed-off-by: Allain Legacy <alegacy@redhat.com>