Intercept and convert plain text errors to JSON (#473)
This implements a workaround for a limitation of the http.ServeMux to
convert plain text error responses to JSON.  The limitation is that it
does not provide a means to specify whether error responses should be
in plain text or JSON format.  The issue is discussed here [1].

The 'solution' is to wrap the ServeMux so that we supply a custom
ResponseWriter that is capable of intercepting "text/plain" errors
and re-format them to "application/problem+json" to comply with the
ORAN interface specifications.

The previous approach attempted to better handle this scenario for a
404 error using a default "/" path, but overlooked that 405 errors are
never returned as long as the "/" path is in the mux.

[1] https://github.com/golang/go/issues/65648

Signed-off-by: Allain Legacy <alegacy@redhat.com>