Fix the end of sync query when not data is present (#483)
When the resync ends and no data exists in the k8s API for some resource
then the "NotIn" clause on the database query throws an error since the
list passed to it is empty.  If no data exists in the API server then we
simply want to delete ALL records therefore we don't need to supply a
where clause.

Signed-off-by: Allain Legacy <alegacy@redhat.com>