Require cluster template labels on clusters and agents
This change adds a validation step on events received for Managed
Clusters and Agents.  If they do not have a condition that indicates
they are ready, or they do not have the cluster template id label that
indicates the provisioning request has been fulfilled, then we simply
ignore them until they do.

Signed-off-by: Allain Legacy <alegacy@redhat.com>