Fetch alarm dictionaries from cluster server
The alarm server uses the alarm dictionary endpoints
of the cluster server to fetch the alarm dictionaries.
These dictionaries are cached in the same way as the
nodeClusters and nodeClusterTypes objects.

There are 2 new methods in the infrastructure client to
return the nodeClusterTypeID and the alarmDefinitionID.
These methods return the IDs if they are present in cache
or try to fetch them from the cluster server if they are missing.

There is also a resync mechanism that runs every hour.

Signed-off-by: Marcelo Guerrero <marguerr@redhat.com>