Fix debug mode handling for re-registration
The REGISTER_ON_RESTART env variable was meant to force a single
registration on every Pod restart, but it was inadvertently causing a
registration on every reconcile loop.  It was also noticed that when
deploying from a bundle that it is not straightforward to set an env
variable as it gets overridden automatically to align with the CSV,
therefore, we are changing this to an annotation.

Signed-off-by: Allain Legacy <alegacy@redhat.com>