Change field include and exclude handling
This changes the implementation of the include/exclude field handling
so that it is done explicitly in the API handlers rather than
implicitly in the middleware layer.  It also adds support for the
"all_fields" query parameter.

The response behavior will now be that any optional field that is nil
will be excluded from responses unless the "all_fields" parameter is
set, or the field name appears in the "fields=" query parameter.

Signed-off-by: Allain Legacy <alegacy@redhat.com>