Set unique UUID value for deployment managers (#538)
Currently, both the NodeCluster and DeploymentManager directly use the
ManagedCluster "clusterID" label as their UUID.  This causes a conflict
between the two API namespaces.  Since the NodeCluster needs the raw
UUID so that we can more easily map incoming alerts to it therefore we
will create a unique value for DeploymentManagers to avoid the
collision.

Signed-off-by: Allain Legacy <alegacy@redhat.com>