Fix panic on duplicate node clusters in DB
We never expect that there will be multiple NodeCluster objects in the
database with the same name.  When we query for NodeCluster objects with
a specific name, we are returned an error indicating that multiple rows
were found when only one was expected.  This error is not handled, and
we dereference a nil pointer.

Signed-off-by: Allain Legacy <alegacy@redhat.com>